moved andthen extension, ListK and his extension to arrow-core-data
