+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/ListKTest.kt
@@ -1,19 +1,21 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
-import arrow.core.Tuple2
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
 import arrow.core.extensions.tuple2.eq.eq
-import arrow.data.extensions.listk.applicative.applicative
-import arrow.data.extensions.listk.eq.eq
-import arrow.data.extensions.listk.hash.hash
-import arrow.data.extensions.listk.monoidK.monoidK
-import arrow.data.extensions.listk.monoidal.monoidal
-import arrow.data.extensions.listk.semigroupK.semigroupK
-import arrow.data.extensions.listk.show.show
-import arrow.data.extensions.listk.traverse.traverse
+import arrow.core.ForListK
+import arrow.core.ListKOf
+import arrow.core.extensions.listk.applicative.applicative
+import arrow.core.extensions.listk.eq.eq
+import arrow.core.extensions.listk.hash.hash
+import arrow.core.extensions.listk.monoidK.monoidK
+import arrow.core.extensions.listk.monoidal.monoidal
+import arrow.core.extensions.listk.semigroupK.semigroupK
+import arrow.core.extensions.listk.show.show
+import arrow.core.extensions.listk.traverse.traverse
 import arrow.mtl.extensions.listk.monadCombine.monadCombine
+import arrow.core.fix
 import arrow.test.UnitSpec
 import arrow.test.laws.HashLaws
 import arrow.test.laws.MonadCombineLaws
