+++ modules/core/arrow-core-extensions/src/main/kotlin/arrow/core/extensions/listk.kt
@@ -1,17 +1,17 @@
-package arrow.data.extensions
+package arrow.core.extensions
 
 import arrow.Kind
 import arrow.core.Either
 import arrow.core.Eval
 import arrow.core.Tuple2
-import arrow.data.ForListK
-import arrow.data.ListK
-import arrow.data.ListKOf
-import arrow.data.k
-import arrow.data.extensions.listk.foldable.foldLeft
-import arrow.data.extensions.listk.monad.map
-import arrow.data.extensions.listk.monad.monad
-import arrow.data.fix
+import arrow.core.ForListK
+import arrow.core.ListK
+import arrow.core.ListKOf
+import arrow.core.k
+import arrow.core.extensions.listk.foldable.foldLeft
+import arrow.core.extensions.listk.monad.map
+import arrow.core.extensions.listk.monad.monad
+import arrow.core.fix
 import arrow.extension
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Apply
